Move 'diff' command above 'echo'
Prior to this change, the script outputs the average and maximum
numerical differences between files and afterwards prints a diff of the
files. These diff texts might be very long and make it hard to reach
the relevant difference information. By swapping the lines the crucial
message is always readable, no matter how long the diff output gets.