Print avg/max diff values after file content
Moves the diff command above the line printing avg and max diff.
This is to make it easier to see average/maximal diff values when
the amount of line differences are very large in a file. Otherwise
the 'diff' command might flood the output and make the avg/max
line hard to find.